# base/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Học Tiếng Anh",
        "description": "Tài liệu API cho ứng dụng học tiếng Anh.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lấy thông tin tài khoản người dùng + thống kê điểm",
                "operationId": "695a5ba8ff7a490aad2bed007b5499f9",
                "responses": {
                    "200": {
                        "description": "Thông tin tài khoản và thống kê",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thành công!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nguyễn Văn A"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "a@gmail.com"
                                                },
                                                "total_score": {
                                                    "type": "number",
                                                    "example": 34.5
                                                },
                                                "total_lessons": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "last_lesson": {
                                                    "properties": {
                                                        "lesson_id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "score": {
                                                            "type": "number",
                                                            "example": 9
                                                        },
                                                        "percent": {
                                                            "type": "number",
                                                            "example": 90
                                                        },
                                                        "submitted_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-04-08T22:10:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực người dùng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user-score/add": {
            "post": {
                "tags": [
                    "User Score"
                ],
                "summary": "Lưu điểm người dùng sau khi hoàn thành bài học",
                "operationId": "c24696ec8b06a973f137c98308d67d96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "percent"
                                ],
                                "properties": {
                                    "percent": {
                                        "description": "Phần trăm kết quả đúng (0–100)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 85.5
                                    },
                                    "lesson_id": {
                                        "description": "ID của bài học (nếu có)",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "time_spent": {
                                        "description": "Thời gian học bài (tính bằng giây)",
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lưu điểm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Score saved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "lesson_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "percent": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 85.5
                                                },
                                                "score": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 8.6
                                                },
                                                "submitted_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-08T21:58:00Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu đầu vào không hợp lệ"
                    },
                    "401": {
                        "description": "Chưa xác thực người dùng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/topic/completed": {
            "post": {
                "tags": [
                    "Topic"
                ],
                "summary": "Đánh dấu topic là hoàn thành",
                "operationId": "b66fd6d46f8e101b485c98de6ecef001",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "topic_id",
                                    "level"
                                ],
                                "properties": {
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đánh dấu hoàn thành thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi đầu vào hoặc logic"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}